some placeholders for an alternate method of doing this
